html/browsers/browsing-the-web/back-forward-cache/timers.html

<!doctype html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/cross-origin-embedder-policy/credentialless/resources/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
// https://bugs.chromium.org/p/chromium/issues/detail?id=1202687
// https://html.spec.whatwg.org/multipage/timers-and-user-prompts.html#timers:fully-active
promise_test(async t => {
  const idA = token();
  const idB = token();

  const delayMain = 8000;
  const delayBeforeNavigation = 3000;
  const delayBeforeBackNavigation = 3000;

  window.open(executorPath + idA, '_blank', 'noopener');

  const startTime = performance.now();

  await send(idA, `
    setTimeout(() => send('${idThis}', 'Main timer fired'), ${delayMain});
    setTimeout(() => {
      prepareNavigation();
      location.href = '${originCrossSite + executorPath + idB}';
    }, ${delayBeforeNavigation});
  `);

  send(idB, `
    setTimeout(() => history.back(), ${delayBeforeBackNavigation});
    await send('${idThis}', 'Navigated');
  `);
  assert_equals(await receive(idThis), 'Navigated');

  await assert_bfcached(idA);

  assert_equals(await receive(idThis), 'Main timer fired');

  const actualDelay = performance.now() - startTime;

  // Expected delay:
  //    `delayMain + delayBeforeBackNavigation`
  //    (The timer stops during the page is in BFCache)
  // Failure #1:
  //    `delayMain` (The timer doesn't stop)
  // Failure #2:
  //    `delayMain + delayBeforeBackNavigation + delayBeforeNavigation`
  //    (The timer is reset when restored from BFCache)

  const slack = 2500;
  assert_less_than(slack, delayBeforeNavigation,
    'Slack should be < delayBeforeNavigation to catch Failure #2');

  assert_between_inclusive(
    actualDelay,
    delayMain + delayBeforeBackNavigation,
    delayMain + delayBeforeBackNavigation + slack,
    'Actual delay');

}, 'Timers should stop during the page is in BFCache');
</script>
